Adding an InvalidHost exception class
