+++ src/InvalidDomain.php
@@ -15,38 +15,18 @@ declare(strict_types=1);
 
 namespace Pdp;
 
-use InvalidArgumentException;
 use function sprintf;
 
-class InvalidDomain extends InvalidArgumentException implements ExceptionInterface
+class InvalidDomain extends InvalidHost
 {
     public static function dueToInvalidLabelKey(int $key): self
     {
         return new self(sprintf('the given key `%s` is invalid', $key));
     }
 
-    public static function dueToUnsupportedLabelType(): self
-    {
-        return new self('The label can not be NULL');
-    }
-
     public static function dueToInvalidCharacters(string $domain): self
     {
-        return new self(sprintf('The host `%s` is invalid: it contains invalid characters', $domain));
-    }
-
-    public static function dueToIDNAError(string $domain, string $message = ''): self
-    {
-        if ('' === $message) {
-            return new self(sprintf('The host `%s` is invalid', $domain));
-        }
-
-        return new self(sprintf('The host `%s` is invalid : %s', $domain, $message));
-    }
-
-    public static function dueToMalformedLabels(string $domain): self
-    {
-        return new self(sprintf('The domain `%s` is invalid: the labels are malformed', $domain));
+        return new self(sprintf('The domain `%s` is invalid: it contains invalid characters', $domain));
     }
 
     public static function dueToUnsupportedType(string $domain): self
